==> 1. Splitting based on class.ipynb <==
import tensorflow as tf
import pandas as pd
import numpy as np
import seaborn as sns

import matplotlib.pyplot as plt

df_train = pd.read_csv('./0.train.csv')
df_train.head()

neg, pos = np.bincount(df_train['Best Performance'].to_numpy())
print("negative class :", neg)
print("positive class :", pos)
print("total :", neg + pos)

# Split positive and negative class

by_class = df_train.groupby('Best Performance')

df_train_positive = by_class.get_group(1)
df_train_negative = by_class.get_group(0)

print(df_train_positive.shape)
print(df_train_negative.shape)

df_train_positive.to_csv("1.train_positive.csv", index=False)
df_train_negative.to_csv("1.train_negative.csv", index=False)
