==> 2. Dealing with imbalance data.ipynb <==
import tensorflow as tf
import pandas as pd
import numpy as np
import seaborn as sns

import matplotlib.pyplot as plt

To deal with umbalance data, I performed undersampling and oversampling. I dropped some sample in negative class that has missing data or redundant data (undersampling). Then I replicate sample in positive class (oversampling). Thus, the amount of sample between positive and negative data is nearly the same.

df_train_pos = pd.read_csv('./1.train_positive.csv')
df_train_neg = pd.read_csv('./1.train_negative.csv')

print(df_train_pos.shape)
print(df_train_neg.shape)

# Check duplicate row from negative class and drop it

# Select duplicate rows except first occurrence based on all columns
duplicateRowsDF = df_train_pos[df_train_pos.duplicated()]
print("Duplicate Rows except first occurrence based on all columns are :")
print(duplicateRowsDF)

# Select duplicate rows except first occurrence based on all columns
duplicateRowsDF = df_train_neg[df_train_neg.duplicated()]
print("Duplicate Rows except first occurrence based on all columns are :")
print(duplicateRowsDF)

# df_train_neg.drop(df_train_neg.loc[df_train_neg['line_race']==0].index, inplace=True)
# dropping duplicate values 
df_train_neg.drop_duplicates(keep=False,inplace=True) 
print(df_train_neg.shape)

# Check missing row from negative class and drop it

null_col = df_train_neg.columns[df_train_neg.isna().any()]
null_df = df_train_neg[null_col].isna().sum().rename('missing rows').to_frame()
# null_df['percentage'] = round(null_df['missing rows'] / df_train.shape[0] * 100, 3)
# null_df['percentage'] = null_df['percentage'].astype('str')

null_df.sort_values('missing rows', ascending=False).style.background_gradient('Blues')

# Remove row having missing data

print(df_train_neg.shape)

# drop rows with missing values
df_train_neg.dropna(inplace=True)

print(df_train_neg.shape)

# replicate row in positive class

df_train_pos = pd.concat([df_train_pos]*6)

print(df_train_pos.shape)

# Merge positive and negative data frame

new_df_train = pd.concat([df_train_pos,df_train_neg])
print(new_df_train.shape)

new_df_train.to_csv("2.balance_train.csv", index=False)
