==> 3. Convert all data into numerical data.ipynb <==
import tensorflow as tf
import pandas as pd
import numpy as np
import seaborn as sns
from numpy import nan

import matplotlib.pyplot as plt

df_train = pd.read_csv('./2.balance_train.csv')
df_test = pd.read_csv('./0.test.csv')

print(df_train.shape)
print(df_test.shape)

df_train.dtypes

There are 21 attributes with different types of value such as int, float, and object. 
We only convert the object type to int types.

# job_level                                 

print(df_train['job_level'].unique())
print(df_test['job_level'].unique())

df_train['job_level'] = df_train['job_level'].replace("JG05", 5)
df_train['job_level'] = df_train['job_level'].replace("JG04", 4)
df_train['job_level'] = df_train['job_level'].replace("JG03", 3)
df_train['job_level'] = df_train['job_level'].replace("JG06", 6)

df_test['job_level'] = df_test['job_level'].replace("JG04", 4)
df_test['job_level'] = df_test['job_level'].replace("JG03", 3)
df_test['job_level'] = df_test['job_level'].replace("JG05", 5)
df_test['job_level'] = df_test['job_level'].replace("JG06", 6)

print(df_train['job_level'].unique())
print(df_test['job_level'].unique())

# person_level                              

print(df_train['person_level'].unique())
print(df_test['person_level'].unique())

for i in range(8):
    temp='PG0'+str(i+1)
    df_train['person_level'] = df_train['person_level'].replace(temp, (i+1))
    df_test['person_level'] = df_test['person_level'].replace(temp, (i+1))
    
print(df_train['person_level'].unique())
print(df_test['person_level'].unique())

# Employee_type                             

print(df_train['Employee_type'].unique())
print(df_test['Employee_type'].unique())

df_train['Employee_type'] = df_train['Employee_type'].replace("RM_type_A", 1)
df_train['Employee_type'] = df_train['Employee_type'].replace("RM_type_B", 2)
df_train['Employee_type'] = df_train['Employee_type'].replace("RM_type_C", 3)

df_test['Employee_type'] = df_test['Employee_type'].replace("RM_type_A", 1)
df_test['Employee_type'] = df_test['Employee_type'].replace("RM_type_B", 2)
df_test['Employee_type'] = df_test['Employee_type'].replace("RM_type_C", 3)

print(df_train['Employee_type'].unique())
print(df_test['Employee_type'].unique())

df_train['gender'] = df_train['gender'].replace("Male", 1)
df_train['gender'] = df_train['gender'].replace("Female", 0)

df_test['gender'] = df_test['gender'].replace("Male", 1)
df_test['gender'] = df_test['gender'].replace("Female", 0)

print(df_train['gender'].unique())
print(df_test['gender'].unique())

# marital_status_maried(Y/N)

print(df_train['marital_status_maried(Y/N)'].unique())
print(df_test['marital_status_maried(Y/N)'].unique())

df_train['marital_status_maried(Y/N)'] = df_train['marital_status_maried(Y/N)'].replace("Y", 1)
df_train['marital_status_maried(Y/N)'] = df_train['marital_status_maried(Y/N)'].replace("N", 0)

df_test['marital_status_maried(Y/N)'] = df_test['marital_status_maried(Y/N)'].replace("Y", 1)
df_test['marital_status_maried(Y/N)'] = df_test['marital_status_maried(Y/N)'].replace("N", 0)

print(df_train['marital_status_maried(Y/N)'].unique())
print(df_test['marital_status_maried(Y/N)'].unique())

# Education_level                           

print(df_train['Education_level'].unique())
print(df_test['Education_level'].unique())

def convert_edu_level(x):
    try:
        if x=="level_0":
            return 0
        elif x=="level_1":
            return 1
        elif x=="level_2":
            return 2
        elif x=="level_3":
            return 3
        elif x=="level_4":
            return 4
        elif x=="level_5":
            return 5
    except:
        return nan

df_train['Education_level'] = df_train['Education_level'].apply(convert_edu_level)
df_test['Education_level'] = df_test['Education_level'].apply(convert_edu_level)

print(df_train['Education_level'].unique())
print(df_test['Education_level'].unique())

# year_graduated                            

print(df_train['year_graduated'].unique())
print(df_test['year_graduated'].unique())

def filter_year(x):
    try:
        x = int(x)
        if x > 2021:
            return nan #make invalid data into nan/empty instead
        elif x < 1945:
            return nan #make invalid data into nan/empty instead
        elif x > 1945 and x<=2021:
            return 2021 - int(x)
    except:
        return nan

df_train['year_graduated'] = df_train['year_graduated'].apply(filter_year)
df_test['year_graduated'] = df_test['year_graduated'].apply(filter_year)

print(df_train['year_graduated'].unique())
print(df_test['year_graduated'].unique())

# age

print(df_train['age'].unique())
print(df_test['age'].unique())

def filter_age(x):
    try:
        x = int(x)
        if x > 2004:
            return nan #make invalid data into nan/empty instead
        elif x < 1945:
            return nan #make invalid data into nan/empty instead
        elif x > 1945 and x<=2004:
            return 2021 - int(x)
    except:
        return nan

df_train['age'] = df_train['age'].apply(filter_age)
df_test['age'] = df_test['age'].apply(filter_age)

print(df_train['age'].unique())
print(df_test['age'].unique())

df_train.to_csv("3.num_train.csv", index=False)
df_test.to_csv("3.num_test.csv", index=False)
