==> 5. Dealing with missing data.ipynb <==
import tensorflow as tf
import pandas as pd
import numpy as np
import seaborn as sns

import matplotlib.pyplot as plt

I transformed the missing data (nan) with mean (for real type data) and median (for integer type data)

df_train = pd.read_csv('./4.train_without_outlier.csv')
df_test = pd.read_csv('./4.test_without_outlier.csv')

print(df_train.shape)
print(df_test.shape)

df_merge_data = pd.concat([df_train,df_test])

df_merge_data.dtypes

null_col = df_merge_data.columns[df_merge_data.isna().any()]
null_df = df_merge_data[null_col].isna().sum().rename('missing rows').to_frame()
null_df['percentage'] = round(null_df['missing rows'] / df_merge_data.shape[0] * 100, 3)
null_df['percentage'] = null_df['percentage'].astype('str')

null_df.sort_values('missing rows', ascending=False).style.background_gradient('Blues')

# GPA

print(df_merge_data['GPA'].unique())

# fill missing values with mean column values
df_merge_data['GPA'].fillna(np.round(df_merge_data['GPA'].mean(),2), inplace=True)
print(df_merge_data['GPA'].unique())

# Achievement_above_100%_during3quartal

# count the number of NaN values in each column
print(df_merge_data['Achievement_above_100%_during3quartal'].isna().sum())

print(df_merge_data['Achievement_above_100%_during3quartal'].unique())

# fill missing values with median column values
df_merge_data['Achievement_above_100%_during3quartal'].fillna(df_merge_data['Achievement_above_100%_during3quartal'].median(), inplace=True)
print(df_merge_data['Achievement_above_100%_during3quartal'].unique())

# assign_of_otherposition

print(df_merge_data['assign_of_otherposition'].unique())

# fill missing values with median column values
df_merge_data['assign_of_otherposition'].fillna(df_merge_data['assign_of_otherposition'].median(), inplace=True)
print(df_merge_data['assign_of_otherposition'].unique())

# job_duration_in_current_person_level

print(df_merge_data['job_duration_in_current_person_level'].unique())

# fill missing values with mean column values
df_merge_data['job_duration_in_current_person_level'].fillna(np.round(df_merge_data['job_duration_in_current_person_level'].mean(),2), inplace=True)
print(df_merge_data['job_duration_in_current_person_level'].unique())

# sick_leaves

print(df_merge_data['sick_leaves'].unique())

# fill missing values with median column values
df_merge_data['sick_leaves'].fillna(df_merge_data['sick_leaves'].median(), inplace=True)
print(df_merge_data['sick_leaves'].unique())

# job_duration_in_current_job_level

print(df_merge_data['job_duration_in_current_job_level'].unique())

# fill missing values with mean column values
df_merge_data['job_duration_in_current_job_level'].fillna(np.round(df_merge_data['job_duration_in_current_job_level'].mean(),2), inplace=True)
print(df_merge_data['job_duration_in_current_job_level'].unique())

# job_duration_from_training

print(df_merge_data['job_duration_from_training'].unique())

# fill missing values with median column values
df_merge_data['job_duration_from_training'].fillna(df_merge_data['job_duration_from_training'].median(), inplace=True)
print(df_merge_data['job_duration_from_training'].unique())

# year_graduated

print(df_merge_data['year_graduated'].unique())

# fill missing values with median column values
df_merge_data['year_graduated'].fillna(df_merge_data['year_graduated'].median(), inplace=True)
print(df_merge_data['year_graduated'].unique())

# branch_rotation

print(df_merge_data['branch_rotation'].unique())

# fill missing values with median column values
df_merge_data['branch_rotation'].fillna(df_merge_data['branch_rotation'].median(), inplace=True)
print(df_merge_data['branch_rotation'].unique())

# job_rotation

print(df_merge_data['job_rotation'].unique())

# fill missing values with median column values
df_merge_data['job_rotation'].fillna(df_merge_data['job_rotation'].median(), inplace=True)
print(df_merge_data['job_rotation'].unique())

# job_duration_in_current_branch

print(df_merge_data['job_duration_in_current_branch'].unique())

# fill missing values with mean column values
df_merge_data['job_duration_in_current_branch'].fillna(np.round(df_merge_data['job_duration_in_current_branch'].mean(),2), inplace=True)
print(df_merge_data['job_duration_in_current_branch'].unique())

# annual leave

print(df_merge_data['annual leave'].unique())

# fill missing values with median column values
df_merge_data['annual leave'].fillna(df_merge_data['annual leave'].median(), inplace=True)
print(df_merge_data['annual leave'].unique())

# number_of_dependences

print(df_merge_data['number_of_dependences'].unique())

# fill missing values with median column values
df_merge_data['number_of_dependences'].fillna(df_merge_data['number_of_dependences'].median(), inplace=True)
print(df_merge_data['number_of_dependences'].unique())

# age

print(df_merge_data['age'].unique())

# fill missing values with median column values
df_merge_data['age'].fillna(df_merge_data['age'].median(), inplace=True)
print(df_merge_data['age'].unique())

df_train_new = df_merge_data.iloc[:len(df_train),:] 
df_test_new = df_merge_data.iloc[len(df_train):,:]

print(df_train_new.shape)
print(df_test_new.shape)

df_train_new.to_csv("5.complete_train.csv", index=False)
df_test_new.to_csv("5.complete_test.csv", index=False)
