==> 9. Implementing PCA.ipynb <==
import pandas as pd 
import numpy as np
from numpy import nan
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA

df_train = pd.read_csv("7.selectedfeature_train.csv") 
df_test = pd.read_csv("7.selectedfeature_test.csv") 

print(df_train.shape)
print(df_test.shape)

df_merge_data = pd.concat([df_train,df_test])
print(df_merge_data.shape)

label = df_merge_data["Best Performance"] #label
df_merge_data.drop(['Best Performance'], inplace=True, axis=1)

pca = PCA(0.9)
pca.fit(df_merge_data)

PCA(copy=True, iterated_power='auto', n_components=0.9, random_state=None, svd_solver='auto', tol=0.0, whiten=False)
pca.n_components_

train_pca = pca.transform(df_merge_data)
#karena tadi udah fit, kalo belum fit bisa gini :
#train_pca = pca.fit_transform(data)
train_pca.shape

feat_cols = ['feature'+str(i) for i in range(train_pca.shape[1])]
pca_data = pd.DataFrame(train_pca,columns=feat_cols)

pca_data.head()

label.head()

print(pca_data.shape)
print(label.shape)

pca_data.reset_index(drop=True, inplace=True)
label.reset_index(drop=True, inplace=True)
result = pd.concat([pca_data, label], axis=1)
result.shape

result.head()

df_train_new = result.iloc[:len(df_train),:] 
df_test_new = result.iloc[len(df_train):,:]
print(df_train_new.shape)
print(df_test_new.shape)

df_train_new.to_csv("9.pca_train.csv", index=False)
df_test_new.to_csv("9.pca_test.csv", index=False)
